Move more stuff around and fix surrogate fill-in in Jacobian for sequential block solve
